+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  
 ### Changed
  - Fixed exception trying to get DHCP status of network interfaces on Linux
+ - Fixed a bug where color formatted console output on Net Standard was not raising the OnConsolePrint event
 
 ## [10.3.0] - 2020-01-20
 ### Changed